Disable proxy buffering on nginx.
nginx has, by default, an output buffering. We can turn it off by using
the `proxy_buffering` directive but it can be cumbersome in some cases.
We let the user decide whether adding this directive or not. However, we
use the `X-Accel-Buffering` header to disable the buffering for this
specific resource. Please, see
http://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_buffering.